seprating artistname search process into  new file
